Adds balance get & update APIs (#263)
* build: removes coin related logic from exchange principal cloud function

* build: adds mechanism to fetch balance from backend & update it to firebase at app launch

* chore: fixes state bug in leaderboard

* chore: removes unwanted files

* build: adds JWT to Firebase secrets

* chore: fixes firebase signout issue during Apple/Google login

* build: adds extension to use hex code for colours

* chore: fixes colour extensions

* chore: fixes smiley game result now show if scrolled before animation success

* chore: fixes error view for smiley game

* chore: uses hex instead of colour string for game rules

* build: adds firebase prod to config

* build: adds seed vote logic for new videos

* chore: fixes routing from profile to upload view

* build: adds logic to support dynamic batch size for feeds

* chore: adds sats balance base url to app configuration

* chore: adds library search paths for yral

+++ iosApp/iosApp/Libs/Feed/View/Helper/ProfileInfoView.swift
@@ -115,7 +115,7 @@ class ProfileInfoView: UIView, ImageLoaderProtocol {
     let imageURL: URL?
     let title: String
     let subtitle: String
-    var coins: Int
+    var coins: UInt64
   }
 }
